+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add measure_time context for measuring exection time (#765)
 - Add method to combine legends of multiple matplotlib axes (#782)
+- Functionality plot onto a matplotlib Axes object (#785)
+
 ### Changed
 
 - Improve sine fitting (#762)
@@ -77,7 +79,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update docstrings (#718)
 - Add numdifftools to requirements (#726)
 - Add fitting of sine wave (#731)
-- Functionality plot onto a matplotlib Axes object (#785)
 
 ### Changed
 - Prevent tests from writing data to the user directory (#710)